+++ components/post_view/reaction_list/index.ts
@@ -5,15 +5,17 @@ import {connect} from 'react-redux';
 import {bindActionCreators, Dispatch} from 'redux';
 
 import {getChannel} from 'mattermost-redux/selectors/entities/channels';
-import {makeGetReactionsForPost} from 'mattermost-redux/selectors/entities/posts';
 import {canAddReactions} from 'mattermost-redux/selectors/entities/reactions';
 
 import {GenericAction} from 'mattermost-redux/types/actions';
 import {Post} from 'mattermost-redux/types/posts';
 import {Reaction} from 'mattermost-redux/types/reactions';
-import {GlobalState} from 'mattermost-redux/types/store';
+
+import {GlobalState} from 'types/store';
 
 import {addReaction} from 'actions/post_actions.jsx';
+
+import {makeGetUniqueReactionsToPost} from 'utils/post_utils';
 
 import ReactionList from './reaction_list';
 
@@ -25,7 +27,7 @@ type Props = {
 };
 
 function makeMapStateToProps() {
-    const getReactionsForPost = makeGetReactionsForPost();
+    const getReactionsForPost = makeGetUniqueReactionsToPost();
 
     return function mapStateToProps(state: GlobalState, ownProps: Props) {
         const channelId = ownProps.post.channel_id;
