+++ packages/mattermost-redux/src/selectors/entities/posts.ts
@@ -41,8 +41,6 @@ import {
 } from 'mattermost-redux/utils/post_utils';
 import {getPreferenceKey} from 'mattermost-redux/utils/preference_utils';
 
-import {getEmojiMap} from 'selectors/emojis';
-
 export function getAllPosts(state: GlobalState) {
     return state.entities.posts.posts;
 }
@@ -65,27 +63,16 @@ export function getReactionsForPosts(state: GlobalState): RelationOneToOne<Post,
     return state.entities.posts.reactions;
 }
 
-export function makeGetReactionsForPost(): (state: GlobalState, postId: Post['id']) => Record<string, Reaction> | undefined | null {
-    return createSelector(
-        'makeGetReactionsForPost',
-        getEmojiMap,
-        getReactionsForPosts,
-        (_: GlobalState, postId: string) => postId,
-        (emojiMap, reactions, postId) => {
-            if (!reactions[postId]) {
-                return null;
-            }
+export function makeGetReactionsForPost(): (state: GlobalState, postId: $ID<Post>) => {
+    [x: string]: Reaction;
+} | undefined | null {
+    return createSelector('makeGetReactionsForPost', getReactionsForPosts, (state: GlobalState, postId: string) => postId, (reactions, postId) => {
+        if (reactions[postId]) {
+            return reactions[postId];
+        }
 
-            const reactionsForPost: Record<string, Reaction> = {};
-
-            Object.entries(reactions[postId]).forEach(([userIdEmojiKey, emojiReaction]) => {
-                if (emojiMap.get(emojiReaction.emoji_name)) {
-                    reactionsForPost[userIdEmojiKey] = emojiReaction;
-                }
-            });
-
-            return reactionsForPost;
-        });
+        return null;
+    });
 }
 
 export function getOpenGraphMetadata(state: GlobalState): RelationOneToOne<Post, Dictionary<OpenGraphMetadata>> {
@@ -388,7 +375,9 @@ export function makeGetProfilesForThread(): (state: GlobalState, rootId: string)
             const profileIds = posts.map((post) => post.user_id);
             const uniqueIds = [...new Set(profileIds)];
             return uniqueIds.reduce((acc: UserProfile[], id: string) => {
-                const profile: UserProfile = userStatuses ? {...allUsers[id], status: userStatuses[id]} : {...allUsers[id]};
+                const profile: UserProfile = userStatuses ?
+                    {...allUsers[id], status: userStatuses[id]} :
+                    {...allUsers[id]};
 
                 if (profile && Object.keys(profile).length > 0 && currentUserId !== id) {
                     return [
